+++ src/main/kotlin/com/loadingbyte/cinecred/project/Drawn.kt
@@ -13,7 +13,17 @@ class DrawnPage(val defImage: DeferredImage, val stageInfo: PersistentList<Drawn
 
 
 sealed class DrawnStageInfo {
-    class Card(val middleY: Y) : DrawnStageInfo()
-    class Scroll(val scrollStartY: Y, val ownedScrollHeight: Y, val frames: Int, val initialAdvance: Double) :
-        DrawnStageInfo()
+
+    class Card(
+        val middleY: Y
+    ) : DrawnStageInfo()
+
+    class Scroll(
+        val scrollStartY: Y,
+        val scrollStopY: Y,
+        val ownedScrollHeight: Y,
+        val frames: Int,
+        val initialAdvance: Double
+    ) : DrawnStageInfo()
+
 }