+++ src/main/kotlin/com/loadingbyte/cinecred/ui/DeliveryPanel.kt
@@ -7,9 +7,7 @@ import com.loadingbyte.cinecred.ui.helper.PLAY_ICON
 import net.miginfocom.layout.PlatformDefaults
 import net.miginfocom.layout.UnitValue
 import net.miginfocom.swing.MigLayout
-import javax.swing.JButton
-import javax.swing.JPanel
-import javax.swing.JToggleButton
+import javax.swing.*
 
 
 class DeliveryPanel(ctrl: ProjectController) : JPanel() {
@@ -17,10 +15,21 @@ class DeliveryPanel(ctrl: ProjectController) : JPanel() {
     val configurationForm = DeliverConfigurationForm(ctrl)
     val renderQueuePanel = DeliverRenderQueuePanel(ctrl)
 
+    val specsLabels = List(4) { JLabel() }
+    private val specsAndIssuesPanel = JPanel(MigLayout("insets 0, wrap"))
     val addButton = JButton(l10n("ui.delivery.addToRenderQueue"), ADD_ICON)
     val processButton = JToggleButton(l10n("ui.delivery.processRenderQueue"), PLAY_ICON)
 
     init {
+        val specsPanel = JPanel(MigLayout("insets 1, gap 1"))
+        specsPanel.background = UIManager.getColor("Component.borderColor")
+        for (label in specsLabels) {
+            label.border = BorderFactory.createEmptyBorder(5, 15, 5, 15)
+            label.isOpaque = true
+            specsPanel.add(label)
+        }
+        specsAndIssuesPanel.add(specsPanel, "center")
+
         addButton.addActionListener { configurationForm.addRenderJobToQueue() }
 
         processButton.addItemListener {
@@ -37,11 +46,23 @@ class DeliveryPanel(ctrl: ProjectController) : JPanel() {
             "${v.value}lpx"
         } + ")"
 
-        layout = MigLayout("insets dialog, wrap", "", "[]16[][]")
+        layout = MigLayout("insets dialog, wrap", "", "[]20[]20[][]")
         add(configurationForm, "width (100% - $horInsets)!")
+        add(specsAndIssuesPanel, "center")
         add(addButton, "split 2, growx")
         add(processButton, "growx")
         add(renderQueuePanel, "grow, push")
+    }
+
+    fun clearIssues() {
+        while (specsAndIssuesPanel.componentCount > 1)
+            specsAndIssuesPanel.remove(1)
+        specsAndIssuesPanel.revalidate()
+    }
+
+    fun addIssue(icon: Icon, msg: String) {
+        specsAndIssuesPanel.add(JLabel(msg, icon, JLabel.LEADING))
+        specsAndIssuesPanel.revalidate()
     }
 
 }